repl: pretty print the scheme
