+++ src/lib.rs
@@ -11,5 +11,6 @@ pub mod parse;
 pub mod pretty;
 pub mod syntax;
 pub mod types;
+pub mod util;
 
 pub mod test;
